+++ src/App.js
@@ -1,15 +1,11 @@
 import React from 'react';
 import './App.css';
-import PropTypes from 'prop-types';
 import {
-  BrowserRouter, Route, Link, Switch, NavLink,
+  BrowserRouter, Route, Switch, NavLink,
 } from 'react-router-dom';
 
-const initialTabs = [
-  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
-  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
-  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
-];
+import HomePage from './components/HomePage';
+import TabsPage from './components/TabsPage';
 
 const App = () => (
   <BrowserRouter>
@@ -43,47 +39,5 @@ const App = () => (
     </div>
   </BrowserRouter>
 );
-
-const HomePage = () => <h1>Home page</h1>;
-
-const TabsPage = ({ match }) => (
-  <Tabs
-    tabs={initialTabs}
-    tabId={match.params.tabId}
-  />
-);
-
-const Tabs = ({ tabs, tabId }) => (
-  <>
-    <h2>Tabs</h2>
-    <ul className="tabs">
-      {tabs.map(tab => (
-        <li key={tab.id} className="tab">
-          <Link to={`/tabs/${tab.id}`} className="tab__link">{tab.title}</Link>
-        </li>
-      ))}
-    </ul>
-    <p>
-      {tabId ? tabs.find(tab => tab.id === tabId).content : ''}
-    </p>
-  </>
-);
-
-Tabs.propTypes = {
-  tabs: PropTypes.arrayOf(PropTypes.object).isRequired,
-  tabId: PropTypes.string,
-};
-
-Tabs.defaultProps = {
-  tabId: null,
-};
-
-TabsPage.propTypes = {
-  match: PropTypes.shape({
-    params: PropTypes.shape({
-      tabId: PropTypes.string,
-    }),
-  }).isRequired,
-};
 
 export default App;
