+++ src/components/HomePage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const HomePage = () => <h1>Home page</h1>;
+
+export default HomePage;
